RAC3: Refactor: Extract game interface classes
